IDVA6-709 - simplifying controller code by adding middleware for relevant controllers
